Encapsulate world boot concerns into its own static class methods
