+++ world/CMakeLists.txt
@@ -31,6 +31,7 @@ SET(world_sources
 	world_console_connection.cpp
 	world_server_command_handler.cpp
 	worlddb.cpp
+	world_boot.cpp
 	world_store.cpp
 	zonelist.cpp
 	zoneserver.cpp
@@ -68,6 +69,7 @@ SET(world_headers
 	world_tcp_connection.h
 	world_server_command_handler.h
 	worlddb.h
+	world_boot.h
 	world_event_scheduler.h
 	world_store.h
 	zonelist.h
